PPT Picture Support
